ifrs9-platform/pyproject.toml
[project]
name = "ifrs9-platform"
version = "0.1.0"
description = "IFRS 9 Automation Platform for Commercial Banks in Uganda"
authors = [{name = "IFRS9 Team"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "redis>=5.0.0",
    "pika>=1.3.2",
    "minio>=7.2.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "python-dateutil>=2.8.2",
    "reportlab>=4.0.0",
    "openpyxl>=3.1.0",
    "apscheduler>=3.10.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.0",
    "black>=23.11.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pylint.messages_control]
max-line-length = 100
